Feat/data from ul

+++ api/libs/common/src/database/seeds/index.ts
@@ -1,4 +1,5 @@
 import * as Prisma from '@prisma/client';
+import { insertUlData } from './ulDataSeed';
 import { createLanguageCodes } from './languages';
 import { createUniversities } from './universities';
 import { createUsers } from './users';
@@ -8,37 +9,61 @@ import { createInterests } from './interests';
 import { createProfiles } from './profiles';
 import { createLearningObjectives } from './objective';
 import { createReportCategories } from './report';
+import { parseArgs } from 'node:util';
 
 const prisma = new Prisma.PrismaClient();
 
 const load = async () => {
   try {
-    await prisma.learningLanguages.deleteMany();
-    await prisma.masteredLanguages.deleteMany();
-    await prisma.users.deleteMany();
-    await prisma.refusedTandems.deleteMany();
-    await prisma.organizations.deleteMany();
-    await prisma.translations.deleteMany();
-    await prisma.languageCodes.deleteMany();
-    await prisma.countryCodes.deleteMany();
-    await prisma.proficiencyTests.deleteMany();
-    await prisma.interestCategories.deleteMany();
-    await prisma.interests.deleteMany();
-    await prisma.learningObjectives.deleteMany();
-    await prisma.textContent.deleteMany();
-    await prisma.reportCategories.deleteMany();
-    await prisma.routineExecutions.deleteMany();
-    await prisma.tandems.deleteMany();
+    const { values } = parseArgs({
+      options: {
+        delete: { type: 'boolean' },
+        seed: { type: 'string' },
+      },
+    });
+
+    const deleteCurrent = !!values.delete;
+    const seedRandomDataset = values.seed === 'random';
+    const seedULDataset = values.seed === 'ul';
+
+    if (deleteCurrent) {
+      console.info('[DB seed] delete existing data');
+      await prisma.learningLanguages.deleteMany();
+      await prisma.masteredLanguages.deleteMany();
+      await prisma.users.deleteMany();
+      await prisma.refusedTandems.deleteMany();
+      await prisma.organizations.deleteMany();
+      await prisma.translations.deleteMany();
+      await prisma.languageCodes.deleteMany();
+      await prisma.countryCodes.deleteMany();
+      await prisma.proficiencyTests.deleteMany();
+      await prisma.interestCategories.deleteMany();
+      await prisma.interests.deleteMany();
+      await prisma.learningObjectives.deleteMany();
+      await prisma.textContent.deleteMany();
+      await prisma.reportCategories.deleteMany();
+      await prisma.routineExecutions.deleteMany();
+      await prisma.tandems.deleteMany();
+    }
 
     await createCountries(prisma);
     await createLanguageCodes(prisma);
-    await createUniversities(prisma);
     await createReportCategories(prisma);
     await createProficiencyTests(prisma);
     await createInterests(prisma);
     await createLearningObjectives(prisma);
-    await createUsers(200, 100, prisma);
-    await createProfiles(prisma);
+
+    if (seedULDataset || seedRandomDataset) {
+      await createUniversities(prisma);
+    }
+    if (seedULDataset) {
+      console.info('[DB seed] UL dataset');
+      await insertUlData(prisma);
+    } else if (seedRandomDataset) {
+      console.info('[DB seed] random dataset');
+      await createUsers(200, 100, prisma);
+      await createProfiles(prisma);
+    }
   } catch (e) {
     console.error(e);
     process.exit(1);
